Removed date property for user until further notice, fixed datatype mismatch in getting user ID from API

+++ src/ApiServer/src/main/java/api/repositories/UserRepository.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserRepository extends CrudRepository<User, Long>{
 
-    User findOneByUserId(Integer id);
+    User findOneByUserId(Long id);
 }
